requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-environ==0.9.0
django-modeltranslation==0.18.4
django-rosetta==0.9.8
idna==3.3
polib==1.1.1
pycodestyle==2.9.1
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.12
